tests/wpt/web-platform-tests/html/webappapis/scripting/events/body-exposed-window-event-handlers.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<body></body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var elements = ['body', 'frameset'];
var handlers = [
    'load', 'resize', 'blur', 'focus', 'scroll',
    'afterprint', 'beforeprint', 'beforeunload', 'hashchange',
    'languagechange', 'message', 'offline', 'online', 'pagehide',
    'pageshow', 'popstate', 'rejectionhandled', 'storage',
    'unhandledrejection', 'unload'];

elements.forEach(function (elementName) {
  handlers.forEach(function (eventName) {
    var handlerName = "on" + eventName;

    test(function() {
      var windowHandler = function () { return "Handler attached to the window"; };
      window[handlerName] = windowHandler;

      var d = (new DOMParser).parseFromString('', 'text/html');
      var b = d.createElement(elementName);

      assert_equals(b[handlerName], null);

      window[handlerName] = null;
    }, "Return null when getting the " + eventName + " event handler of a windowless " + elementName);

    test(function() {
      var windowHandler = function () { return "Handler attached to the window"; };
      window[handlerName] = windowHandler;

      var d = (new DOMParser).parseFromString('', 'text/html');
      var b = d.createElement(elementName);
      b[handlerName] = function() { return "Handler attached to windowless element"; };

      assert_equals(window[handlerName], windowHandler);
      assert_equals(b[handlerName], null);

      // Clean up window event handler
      window[handlerName] = null;
    }, "Ignore setting of " + eventName + " window event handlers on windowless " + elementName);
  });
});
</script>
